first upload
Included in analysis.R are a series of questions we have brainstormed
and then needs to pick out the good questions to answer. Please add
yours to the file!
Github has the ability to sync the codes on your computer,  track any
changes, and share it with other user. If you want to give a try, I can
show you next time.